New docs and refactored tests
